refactor: rename community components directory to make it a private directory
